Create separate jobService
