+++ src/main/java/bio/terra/buffer/app/controller/JobApiController.java
@@ -2,7 +2,7 @@ package bio.terra.buffer.app.controller;
 
 import bio.terra.buffer.generated.controller.JobsApi;
 import bio.terra.buffer.generated.model.JobModel;
-import bio.terra.buffer.service.resource.FlightScheduler;
+import bio.terra.buffer.service.job.JobService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -11,23 +11,23 @@ import static bio.terra.buffer.app.utils.ControllerUtils.jobToResponse;
 
 @Controller
 public class JobApiController implements JobsApi {
-    private final FlightScheduler flightScheduler;
+    private final JobService jobService;
 
     @Autowired
-    JobApiController(FlightScheduler flightSchedule) {
-        this.flightScheduler = flightSchedule;
+    JobApiController(JobService jobService) {
+        this.jobService = jobService;
     }
 
     @Override
     public ResponseEntity<JobModel> retrieveJob(String id) {
-        JobModel job = flightScheduler.retrieveJob(id);
+        JobModel job = jobService.retrieveJob(id);
         return jobToResponse(job);
     }
 
     @Override
     public ResponseEntity<Object> retrieveJobResult(String id) {
-        FlightScheduler.JobResultWithStatus<Object> resultHolder =
-                flightScheduler.retrieveJobResult(id, Object.class);
+        JobService.JobResultWithStatus<Object> resultHolder =
+                jobService.retrieveJobResult(id, Object.class);
         return ResponseEntity.status(resultHolder.getStatusCode()).body(resultHolder.getResult());
     }
 }
