+++ src/main/java/bio/terra/buffer/app/controller/ResourceApiController.java
@@ -4,6 +4,7 @@ import bio.terra.buffer.common.Pool;
 import bio.terra.buffer.generated.controller.ResourceApi;
 import bio.terra.buffer.generated.model.GoogleProjectUid;
 import bio.terra.buffer.generated.model.JobModel;
+import bio.terra.buffer.service.job.JobService;
 import bio.terra.buffer.service.pool.PoolService;
 import bio.terra.buffer.service.resource.FlightScheduler;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,11 +19,13 @@ import static bio.terra.buffer.app.utils.ControllerUtils.jobToResponse;
 public class ResourceApiController implements ResourceApi {
     private final PoolService poolService;
     private final FlightScheduler flightScheduler;
+    private final JobService jobService;
 
     @Autowired
-    ResourceApiController(PoolService poolService, FlightScheduler flightScheduler) {
+    ResourceApiController(PoolService poolService, FlightScheduler flightScheduler, JobService jobService) {
         this.poolService = poolService;
         this.flightScheduler = flightScheduler;
+        this.jobService = jobService;
     }
 
     @Override
@@ -30,6 +33,6 @@ public class ResourceApiController implements ResourceApi {
         GoogleProjectUid googleProjectUid = new GoogleProjectUid().projectId(projectId);
         Pool pool = poolService.getPoolForGoogleProject(googleProjectUid);
         Optional<String> flightId = flightScheduler.submitRepairResourceFlight(pool, googleProjectUid);
-        return jobToResponse(flightScheduler.retrieveJob(flightId.get()));
+        return jobToResponse(jobService.retrieveJob(flightId.get()));
     }
 }
